Conflicts:
	charts/ingress-controller/templates/ingress-config.yaml
	libs/api/peach-api/borrower/src/lib/api/peach_api.service.ts
	libs/api/peach-api/borrower/src/lib/borrower.module.ts
	libs/api/peach-api/borrower/src/lib/loan/loan.service.ts

+++ libs/api/peach-api/borrower/src/lib/payment-instrument/payment_instrument.service.ts
@@ -69,26 +69,34 @@ export class PeachPaymentInstrumentsService {
     userId: string,
     accountInfo: ConnectAccountDto,
   ): Promise<void> {
-    const borrower: Borrower = await this.borrowerRepository.findOneBy({
+    const borrower: Borrower | null = await this.borrowerRepository.findOneBy({
       userId,
     });
     this.borrowerValidation.isBorrowerDefined(borrower);
 
-    await this.peachApiService.createPlaidPaymentInstrument({
-      publicToken: accountInfo.publicToken,
-      accountId: accountInfo.accountId,
-      personId: borrower.personId,
-    });
+    if (borrower === null) {
+      throw new BorrowerNotFoundError();
+    }
+
+    await this.peachApiService.createPlaidPaymentInstrument(
+      borrower.personId,
+      accountInfo.accountId,
+      accountInfo.publicToken,
+    );
   }
 
   public async removePaymentInstrument(
     userId: string,
     id: string,
   ): Promise<void> {
-    const borrower: Borrower = await this.borrowerRepository.findOneBy({
+    const borrower: Borrower | null = await this.borrowerRepository.findOneBy({
       userId,
     });
     this.borrowerValidation.isBorrowerDefined(borrower);
+
+    if (borrower === null) {
+      throw new BorrowerNotFoundError();
+    }
 
     await this.peachApiService.deletePaymentInstrument(borrower.personId, id);
   }
